+++ compose/src/main/java/com/yourssu/design/system/compose/component/DoubleTitleTopBar.kt
@@ -1,21 +1,17 @@
-package com.yourssu.design.system.compose.components
+package com.yourssu.design.system.compose.component
 
 import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.RowScope
-import androidx.compose.foundation.layout.fillMaxHeight
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.wrapContentHeight
-import androidx.compose.foundation.layout.wrapContentSize
-import androidx.compose.material.Text
 import androidx.compose.material.TopAppBar
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import com.yourssu.design.system.compose.R
@@ -25,38 +21,49 @@ import com.yourssu.design.system.compose.base.YdsScaffold
 import com.yourssu.design.system.compose.base.YdsText
 
 @Composable
-fun SingleTitleTopBar(
-    title: String = "",
+fun DoubleTitleTopBar(
     modifier: Modifier = Modifier,
+    title: String = "",
+    subtitle: String = "",
     actions: @Composable RowScope.() -> Unit = {},
 ) {
     TopAppBar(
         modifier = modifier
             .fillMaxWidth()
-            .height(56.dp),
+            .height(72.dp),
         backgroundColor = YdsTheme.colors.bgElevated,
         contentColor = YdsTheme.colors.textPrimary,
         elevation = 0.dp,
         contentPadding = PaddingValues(end = 4.dp)
     ) {
         Box(
-            modifier = Modifier
-                .fillMaxWidth()
+            modifier = Modifier.fillMaxWidth()
         ) {
-            YdsText(
-                modifier = Modifier
-                    .padding(16.dp, 17.dp, 16.dp, 8.dp)
-                    .wrapContentHeight(),
-                text = title,
-                style = YdsTheme.typography.title2,
-                color = YdsTheme.colors.textPrimary
-            )
+
+            Column(
+                modifier = Modifier.padding(
+                    start = 16.dp,
+                    top = 12.dp,
+                    end = 16.dp,
+                    bottom = 8.dp
+                )
+            ) {
+                YdsText(
+                    text = subtitle,
+                    style = YdsTheme.typography.body2,
+                    color = YdsTheme.colors.textPrimary
+                )
+                YdsText(
+                    text = title,
+                    style = YdsTheme.typography.title2,
+                    color = YdsTheme.colors.textPrimary
+                )
+            }
 
             Row(
                 modifier = Modifier
-                    .fillMaxHeight()
-                    .padding(0.dp)
-                    .align(Alignment.TopEnd),
+                    .padding(top = 16.dp)
+                    .align(Alignment.BottomEnd),
                 verticalAlignment = Alignment.CenterVertically
             ) {
                 actions()
@@ -65,14 +72,15 @@ fun SingleTitleTopBar(
     }
 }
 
-@Preview(name = "SingleTitleTopBar")
+@Preview(name = "DoubleTitleTopBar")
 @Composable
-private fun PreviewSingleTitleTopBar() {
+private fun PreviewDoubleTitleTopBar() {
     YdsTheme {
         YdsScaffold(
             topBar = {
-                SingleTitleTopBar(
+                DoubleTitleTopBar(
                     title = "타이틀",
+                    subtitle = "서브타이틀",
                     actions = {
                         TopBarButton(
                             icon = R.drawable.ic_ground_filled,
