+++ compose/src/main/java/com/yourssu/design/system/compose/component/TopBar.kt
@@ -1,4 +1,4 @@
-package com.yourssu.design.system.compose.components
+package com.yourssu.design.system.compose.component
 
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.PaddingValues
@@ -8,7 +8,6 @@ import androidx.compose.foundation.layout.fillMaxHeight
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.wrapContentSize
 import androidx.compose.material.TopAppBar
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
@@ -23,8 +22,8 @@ import com.yourssu.design.system.compose.base.YdsText
 
 @Composable
 fun TopBar(
-    title: String = "",
     modifier: Modifier = Modifier,
+    title: String = "",
     navigationIcon: @Composable () -> Unit = {},
     actions: @Composable RowScope.() -> Unit = {},
 ) {
